存个档，/emps的后端写好了，来写前端。对我的风格就是先写后端再写掐断。

+++ src/main/java/org/yuliu/heima/demos/web/pojo/PageResult.java
@@ -12,6 +12,6 @@ import java.util.List;
 @AllArgsConstructor
 
 public class PageResult {
-    private int total;
+    private Long total;
     private List rows;
 }
